fastapi/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
alembic
shortuuid
PyJWT